TCES-12 test(auth) add testing for middleware
